tray/requirements.txt
Pillow==9.5.0
PyQRCode==1.2.1
pystray==0.19.3
pywebview==4.3.3
pebble